fuse-wake becomes a subset of wakebox
