+++ Makefile
@@ -45,7 +45,7 @@ clean:
 	rm -f bin/* lib/wake/* */*.o */*/*.o src/json/jlexer.cpp src/parser/lexer.cpp src/parser/parser.cpp src/parser/parser.h src/version.h wake.db
 	touch bin/stamp lib/wake/stamp
 
-wake.db:	bin/wake bin/fuse-wake lib/wake/fuse-waked lib/wake/shim-wake
+wake.db:	bin/wake bin/wakebox lib/wake/fuse-waked lib/wake/shim-wake
 	test -f $@ || ./bin/wake --init .
 
 install:	all
@@ -66,7 +66,7 @@ static:	wake.db
 bin/wake:	$(WAKE_OBJS)
 	$(CXX) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(CORE_LDFLAGS)
 
-bin/fuse-wake:		tools/fuse-wake/fuse-wake.cpp src/wakefs/*.cpp $(COMMON_OBJS)
+bin/wakebox:		tools/wakebox/wakebox.cpp src/wakefs/*.cpp vendor/gopt/*.c $(COMMON_OBJS)
 	$(CXX) $(CFLAGS) $(LOCAL_CFLAGS) $^ -o $@ $(LDFLAGS)
 
 lib/wake/fuse-waked:	tools/fuse-waked/fuse-waked.cpp $(COMMON_OBJS)
