+++ share/wake/lib/system/job.wake
@@ -418,7 +418,7 @@ export def getJobStatus (job: Job): Status = match (getJobReport job)
 #        used relative to where wake has been installed
 export def wakePath: String = prim "execpath" # location of the wake executable
 export def fuseRunner: Runner =
-  def fuse = "{wakePath}/fuse-wake"
+  def fuse = "{wakePath}/wakebox"
   def score plan =
     if plan.getPlanLocalOnly then Fail "would hide workspace" else Pass 1.0
   makeJSONRunnerPlan fuse score
@@ -518,7 +518,7 @@ export def makeJSONRunner (plan: JSONRunnerPlan): Runner =
             Fail f = Pair (Fail (makeError f)) ""
             Pass inFile =
               def outFile = "{build}/{prefix}.out.json"
-              def cmd = script, inFile, outFile, extraArgs
+              def cmd = script, "-p", inFile, "-o", outFile, extraArgs
               def proxy = RunnerInput label cmd Nil (extraEnv ++ environment) "." "" Nil prefix (estimate record)
               Pair (Pass proxy) inFile
   def post = match _
@@ -528,7 +528,7 @@ export def makeJSONRunner (plan: JSONRunnerPlan): Runner =
     Pair (Pass _) inFile =
       def unlink f =
         def fn _ = prim "unlink"
-        match (getenv "FUSE_WAKE_KEEP_IO_FILES")
+        match (getenv "WAKEBOX_KEEP_IO_FILES")
           Some "1" = Unit
           _        = fn f
       def outFile = replace `\.in\.json$` ".out.json" inFile
