--- tools/fuse-wake/fuse-wake.wake
@@ -1,21 +1,0 @@
-# Copyright 2019 SiFive, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You should have received a copy of LICENSE.Apache2 along with
-# this software. If not, you may obtain a copy at
-#
-#    https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-package build_wake
-from wake import _
-from gcc_wake import _
-
-target buildFuse variant: Result (List Path) Error =
-    tool here variant "bin/fuse-wake" (json, wakefs, util, Nil)
