+++ wake.spec.in
@@ -32,7 +32,6 @@ mkdir -p %{buildroot}/var/cache/wake
 %files
 /usr/bin/wake
 /usr/bin/wakebox
-/usr/bin/fuse-wake
 /usr/bin/preload-wake
 /usr/lib/wake
 /usr/share/wake
